# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the forum.",
        "title": "CVWO Forum Backend API",
        "contact": {},
        "license": {
            "name": "All Rights Reserved."
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
        "/comment/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new comment for the given thread",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Handles comment creation requests",
                "parameters": [
                    {
                        "description": "Comment data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    },
                    "400": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Invalid JWT token"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/comment/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates a comment",
                "tags": [
                    "comment"
                ],
                "summary": "Handles comment update requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Invalid JWT token"
                    },
                    "403": {
                        "description": "No permission to update comment"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a comment",
                "tags": [
                    "comment"
                ],
                "summary": "Handles comment deletion requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid JWT token"
                    },
                    "403": {
                        "description": "No permission to delete comment"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/thread/{thread_id}/comments": {
            "get": {
                "description": "Retrieves comments for the given thread",
                "tags": [
                    "comment"
                ],
                "summary": "Handles comment retrieval requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread UUID",
                        "name": "thread_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "created_time_asc",
                            "created_time_desc"
                        ],
                        "type": "string",
                        "description": "Sorting order, default 'created_time_asc'",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number, default '1'",
                        "name": "p",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetCommentResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Comment": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                }
            }
        },
        "models.CreateCommentRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                }
            }
        },
...
        "models.GetCommentResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Comment"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateCommentRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "The word \"Bearer\", followed by a space, and then the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}